Updated: README
Updated the README file to reflect the changes made to the loaders; going from jsx-loader to babel-loader. Have also updated the loaders for the web fonts since issue #24 was raised. This webpack configuration to load web fonts is what I have currently working on 3 projects. Therefore should fix the issue the others were having as well.

closes #24