(docs):Updated README
I have added to the readme a Component API section. This is in the form of a table to show the user the possible props they can pass to the Icon component. Also I have made this table into a separate file, named it README.md and placed it in the src folder. This will give anyone looking at the component an overview of the props.